add: eventHandlers file
This commit adds the eventHandlers file which will be used to handle the interaction logic of the program.
Added a basic click event listener to fire when the add-task-btn is clicked. This calls the renderToDoTask() function which in turn generates a new task form on the page ready for user data entry.