refactor: form creation in renderToDoTask
This commit refactors the code for creating the form in the renderToDoTask() function.

While focusing on a modular approach my previous code had considerable repetition. Upon reflection the following changes have been implemented in this commit.

The refactoring introduces two helper functions: createInputLabel() and createPriorityOption().
These functions encapsulate the creation of the label/input elements and the option elements, reducing code duplication.

I contemplated modifying the createInputLabel function to also have the ability to create the submit button but feel this would be blurring the lines of what the helper function is created to do.