add: addTaskToList function
This commit adds the addTaskToList function which is responsible for handling the to do task form submission and adding a new task to a list.

The plan is to get the basic functionality working so we can display the tasks on screen in a default list.
Additional functionality will be added later, such as for editing task data and moving created tasks between lists.
I am trying to keep a reasonable level of atomic git commits to the project to build it incrementally and logically.