add: renderTaskList function
This commit adds a new function responsible for rendering the task list to the html.
Once a new task has been created and the 'add task' button is clicked, this function is called and renders the task as an li elements and displayed it within the default <ul> list.
At present it only displays the title of each task, to test functionality. This will be iterated on in future commits.