feat: add createDeleteBtn function
This commit adds a new function which renders a delete button element which allows the user to remove a task from the default list.
The renderTaskList function now includes the call to createDeleteBtn to add the delete button to each task item.
The event listener on the delete button triggers the deleteTask function, passing the necessary arguments (index, taskList, renderTaskList) to remove the task and update the rendered task list.