refactor: createTaskItem function
This commit refactors the createTaskItem function to now display all associated data relevant to the task when it is added to the default list.
Now includes the data for properties: description, dueDate, priority.

This allows the user to see all the details of the task in the list rather than just the title.