refactor: createDeleteBtn function
This commit replaces the created button element with a div, this will allow the placement of an trash-can icon within the div instead of a generic button element.
Additionally the text 'delete' has been removed, again for easier css styling purposes later.
To maintain accessability, I have added an aria labelling element via the setAttribute method for assistive technologies.
These changes are to assist with easier styling of the page with CSS.