refactor: to-do task form creation and rendering
The domRendering module was becoming overly complex with functions becoming too tightly coupled.
In an attempt to improve readability and maintainability I have begun to refactor the module.

The changes include creating separate functions for creating to-do form elements (input labels, priority options, and the form itself), rendering the to-do task, and rendering the default list.
Also, the code for handling the list manager's rendering was moved to its own section for better organization.