add: multi-List Functionality in listManager.js
In this commit, the listManager.js module was refactored to accommodate multiple task lists instead of just one. This allows users to manage tasks across multiple lists.

A default list is now added in an array 'allLists', which is used to store all the task lists. The 'createList' function has been updated to create a new list with a specified name and add it to the 'allLists' array.
Each list object consists of a name and an array of tasks.

The 'getAllLists' function has been added to return the 'allLists' array, which can be used to display all the available lists.
The 'deleteTask' function was removed as it will now be handled differently with the introduction of multiple lists.