refactor: use initialise function in index.js
This commit simplifies the set-up of event handlers in the index.js file by using the initialise function which was added to the eventHandlers module.

Previously, individual functions handleAddTaskButtonClick, addTaskToList, and handleAddListButtonClick were imported and called separately in the index.js file.

In this update, all these initialisations are contained within the initialise function. Meaning only this function now needs to be imported and called in index.js, simplifying the code and improving its readability.