refactor: move task addition btn to list container
This commit relocates the 'New Task' button, which is used for adding tasks, to the 'default-list-container' in the index.html file.
Previously, the button was outside the 'default-list-container', in a separate 'add-task-button-container'.

The 'New Task' button is now within the same container as the list to which it adds tasks, making the HTML structure more logical.