feat: add deleteTask functionality for task items
This commit introduces a `deleteTask` function to eventHandlers.

It removes the task at the given index from the taskList and re-renders the task list, updating the user interface accordingly.

This function is to be used in the `renderTaskList` function in the `domRendering` module. When generating each task item, a delete button is created and appended to the item.
An event listener is then added to this button, invoking `deleteTask` when the button is clicked, and removing the task from the list.