feat: update list title dynamically
This commit adds a new function to the domRendering file which updates the default list to display the tasks within a specified list.
The user is now able to click on a list in the list manager, and when selected, the h2 element with the class: 'list-title-text' will be updated dynamically to indicate
to the user which of the lists is being displayed.