feat:add default list to list list manager
Previously the 'default' list which is presented to the user on load was not listed as its own list in the list manager which seemed confusing.

This commit updates the addListToManager function  to check if the list being added is the "default list" (by comparing the listName parameter with the string "default list").
If it is the default list, we insert it at the top of the list manager unordered list using listManagerList.insertBefore(listItem, listManagerList.firstChild).
Otherwise, we add it to the end of the list as before.

This ensures the default list is always first in the list manager, and users can click on each list to view the tasks in the main section.