refactor: createList function to show default list
This commit refactors the createList function with an if statement to check if the listName parameter is equal to 'default list'.
If it is the default list, we simply return the first element of the allLists array, which already contains the default list object with an empty task list.
This ensures that only one default list object exists in the allLists array.