feat: task details expansion and update features
In this commit, I added a new feature that allows users to expand the details of a to-do task for editing.
The details of the task, including the title, description, due date, priority, and list, can now be edited.
When a user clicks on a task in the task list, the task details expand, displaying the existing information in the to-do form.
The submit button in the form is now labeled "update task," and when clicked, it updates the details of the task and re-renders the task list with the changes.

BUG: There is a known bug with this implementation whereby the edited task, will update but then instantly disappear as if the page is being re-rendered.
I need to spend some time resolving this however I have pushed the partially working feature as is so I have a point to refer back to in my commit history.