feat: add modals to get user input
This commit introduces the use of modals to provide a better user experience when adding and editing ToDo tasks.

- Added a new function called openDialog, responsible for creating and displaying the modal dialog.
When this function is called, it creates a dialog element and appends the form elements for creating or updating a ToDo task inside the dialog.
The openDialog function is called when a user clicks on the "Add task" button or when editing a task.
After the task is successfully added or updated, the modal is closed by calling the dialog.close() method.