feat: modals added to event eventHandlers
Following on from the last commit these changes updates the eventhandler file to accommodate the ne modal functionality.

- The handleAddTaskButtonClick function has been modified to use the openDialog function, which was imported from domRendering.js.
When the "Add task" button is clicked, openDialog is called, displaying the modal with the form elements for creating a new task.
Additionally, a new event listener is added to the form inside the modal, listening for a form submission.
When the form is submitted, the addTaskToList function is called to handle the task addition, and then the modal is closed using dialog.close().

- A new function handleEditButtonClick(task) is introduced.
This function is responsible for opening the modal with the form elements filled with the details of the task to be edited.
It works similarly to the handleAddTaskButtonClick function.