feat: add date formatting via date-fns library
This commit adds the date formatting function from the date-fns utility library.
As I am based in the UK the standard date format style is dd/mm/yyyy which has been reflected in the function to format the date rather than the standard js format of: yyyy-mm-dd.

This makes the date format more readable an understandable.