fix: formatDate to handle no selected dates
This commit fixes a bug when the user does not select a date with the date-picker.
Previously when a task was created and no date was selected by the user an error was thrown:
Invalid time value
RangeError: Invalid time value

This is because the dueDate variable isnt assigned an the expected value (a date) and the value is instead set to an empty string which results in the invalid time value error.
To avoid this I've added a conditional ternary operator to return 'null' in cases where the user does not specify a date.
This avoids the error and allows the user to create a task without a specified due date which is a more flexible user experience. The ability to add or change a date is still possible via editing a task by the user.