refactor: add required input for task title
This commit refactors the task creation process to enforce a 'required'attribute on the task title input field, using the native HTML form validation.
This replaces the previous temporary method of checking if the task name was provided during task creation in the createToDoTask function.

The createInputLabel function now accepts an additional optional argument'isRequired' which, when set to true and adds the 'required' attribute to the created input field.
The Title input field in the createToDoForm function has been updated to use this feature, making the task title a required field.

Additionally, the createToDoTask function has been refactored to remove the 'if' check for an empty task name.