fix: triggering of task expansion on delete click
This commit fixes a bug in the code whereby when a user attempts to delete a task by clicking on the delete icon button, the task's details were being expanded.
This happened due to the propagation of the click event from the delete button to the parent task element.

Added a call to event.stopPropagation() in the delete button's click event handler.
This stops the click event from propagating to the parent list item, preventing the task from being expanded when the delete button is clicked.