fix: trigger of task expansion on checkbox click
Similarly to the last fix, this commit addresses a bug in the code whereby when the user clicked on a checkbox to mark a task as complete, the task was expanded to display the details.
This was undesirable as this wouldn't be the expected response by the user.
This happened due to the propagation of the click event from the checkbox icon to the parent task element.

To resolve this issue I have added a call to event.stopPropagation() in the checkbox icon's click event handler.
This stops the click event from propagating to the parent list item, preventing the task from being expanded when the checkbox is clicked.