chore: add console.log()'s for bug tracing
This commit adds some temporary console logs to allow me to track functions and fix some bugs I have in the code.
This is a temporary measure on a new branch and only intended as a reference point to allow me to revert back to the initial code if required when testing.