feat: add cancel btn function
This commit first creates a generic cancel button element named createCancelBtn.
This is then utilised within the openDetailsDialog function to allow the user to close the modal if they decide against making any changes.
This improves the user experience by allowing more flexibility.

The createCancelBtn function will likely be modified at a later date into a div element to hold a suitable icon to represent a cancelling feature but for now a simple button is used for convenience.