refactor: extract date formatting to separate file
This commit moves the date formatting logic into a separate utility module to make the code more maintainable.
This allows for separate control over date formatting rather than it being left in the createToDoTask module which should be useful when I add more date-related utilities are added in the future and need to use it elsewhere in the codebase other than just the createToDoTask function.