fix: dueDate value not being preserved in tasks
Problem: I noticed that when editing a task the dueDate value was reset, this was because the expected date format mismatched with the HTML <input type="date">.

Solution: I have removed the formatting of dueDate in the createToDoTask function, and adapted the code to work with the unformatted date directly.

Before:
dueDate was being formatted using formatDate function in createToDoTask.js.
formatDate was also used in domRendering.js for updating tasks.

After:
Removed the formatting of dueDate from createToDoTask.js.
Adjusted domRendering.js to work directly with the unformatted dueDate.

I will look to implement date formatting at rendering, not in the stored value, to avoid this issue but format the date in a more meaningful way for the user.

All the files/changes related to this fix have been committed together for convenience.