feat: updateTaskDetails to handle task list change
This commit modifies the updateTaskDetails function to allow tasks to be moved between lists when a user updates a task, using the list select input.

Previously, the function updated the task's details but didn't account for possible changes to the task's associated list.
It just added all tasks to the `allTasks` array, without checking that the list associated with a task may have been updated.

In the updated function, I have introduced the following:

1. The function updates the task details as before but now also checks for changes in the task's assigned list.
2. If the task's assigned list has changed, the task is removed from its old list and added to the new list. This allows for tasks to be transferred from one list to another.
3. Lastly, all tasks (from all lists) are pushed onto the `allTasks` array, which is then used to render the task list.