fix: properly assign list name to task object
Previously, the list name was not being passed as an argument to the createToDoTask function during task creation.
This caused the list property of the task object to be undefined, resulting in errors when moving tasks between lists.
This commit fixes the issue by ensuring that the list name is correctly assigned to the task object.