add: setAllLists & Local Storage functionality
A new function setAllLists has been added, which is used in the initialise function to set the allLists array to the data stored in local storage.
This function empties the allLists array and pushes the lists from the stored data into it.

This commit also introduces the data storage functions from the storageModel to the code to save and load/save the list data to local storage.