style: add container for title & btn alignment
This commit adds a div wrapper to allow flex alignment of the title and btn elements on screen.

I've additionally played around with spacing and renaming of elements to make better use of spacing on the page.
This is abit of an experimentation CSS phase as the original design I had planned isn't translating as well as intended so I am granting myself the freedom to make incremental testing adjustments in the following commits.
The following commits will likely be accompanied with less descriptive commit msgs for quicker iteration and experimentation.