refactor: list input rendering
This commit adds a better UX to the creation of lists process.

Previously, the renderListInput function was called during initialisation. This meant the list input section was permanently visible on screen.
This was always a temporary feature to enable quicker testing, however now it has been refactored to allow users to click the add (+) button to create a new list as per my original design intention.

To enable this I have Introduced a new div container, `list-input-container` in the HTML, specifically for holding the list input. This ensures that the existing list manager contents aren't overwritten when rendering the input.
The `renderListInput` function now targets `list-input-container` instead of `list-manager-list`. This avoids overwriting existing list items when showing the input.

Now the user can click on the (+) button, create a new list and when they click 'add list' the input entry field is removed and the + icon needs to be clicked again to create a another list.
I feel this makes the UI more intuitive and less confusing for the user.