refactor: amend naming of create-new-list-btn
This commit renames the create-new-list-btn related elements to simply be new-list-btn. This is more meaningful and simpl as the codebase was becoming unclear.
Having spend some time away from the code I was finding it challenging to follow my previous naming conventions so this is an attempt to streamline and simplify.