style: position buttons to bottom-right corner
This change moves the call to action buttons icons to the bottom right of the container.
The div containers which grouped them together previously with the title have been removed from the HTML as they are no longer required.
The css has been adjusted to utilise the position property to place the buttons into bottom-right corner of the `list-manager-container` and 'default list container'.
I think this gives the layout more space rather than bunching all the elements at the top together making the layout feel cluttered.

Why position instead of Flexbox:
Initially tried to use Flexbox for this adjustment but found that it disrupted the existing layout of other elements within `list-manager-container`.
Additionally the CSS became more confusing due to other flexbox properties being used on these elements as children of their parent container.
Using `position: relative` allowed me to achieve the desired button placement without affecting the layout of sibling elements.

Next:
The + icons are to be changed to simple buttons which will be more intuitive and descriptive with simple "add Task" and "New List" text ensuring their usage is clear rather than ambiguous + icons.