style: consolidate styling and utilise css vars
The css has become somewhat unwieldy so this commit refactors much of the element level applied styles in favour of class based styling to remove unintended side effects.
Still more work to do but this update is part of a few iterations which will simplify the css and make it more manageable by grouping related styles together allowing for universal changes to elements like borders and text.