fix: overlapping btn with list items
This commit fixes a bug I have just discovered.

Issue:
When adding several items to either the list manager or default list, die to using positioning absolute on the button elements, the list items overlap the buttons.
For a temporary fix I have changed the position of the buttons to be top right instead of bottom right, which solves the issue for now but makes the UI a little cluttered.

Given more time I would like to refactor the UI to keep the original position of bottom right for the buttons but this quick fix solves the usability issue for now which is more important.

To compensate I've make the list manager slightly wider to give more spacing between elements.