style: add bg blur on modal interaction
This commit adds some simple JS functions to add and remove a class which is in charge of adding a blur effect to the background when a modal is opened.

This was implemented because without it, the new task and new list buttons are still clearly visible (although not clickable) and were leading to attempted clicks by the user.
This update should improve usability.