fix: set list in task dialog to active list
When a user opens the task creation dialog, the list dropdown select option is now set to the current active list.
This ensures that when the user creates a new task in a list it is automatically added to this list which the user is viewing.
Previously this had to be set by the user manually each time which was rather unintuitive, as users are more likely to expect to add tasks to the currently viewed list.

To achieve this:
- The app now queries the current list title from the list title container.
- Then sets the value of the task-list-select dropdown in the task creation dialog to the fetched list title.